Update
in class

+++ test/project/annotations/AddingMachine.java
@@ -1,9 +1,8 @@
-
 package project.annotations;
 
 public class AddingMachine 
 {
 	public long add(int a, int b) {
-		return a + b;
+		return ((long)a) + b;
 	}
 }
